requirements.txt
fastmcp==0.4.1
websockets==13.0
aiohttp==3.9.5
requests==2.32.3
feedparser==6.0.11
python-dateutil==2.9.0.post0
